Update projects.js
disclaimer: I think this repo might be out of date, so I don't recommend merging this.

I wanted to point out a good opportunity for demonstrating actual coding skills as opposed to demonstrating an HTML wall.  There's other opportunities for this across this repo.  This is by no means necessary, but something to keep in mind.  And if nothing else, good practice!

Instead